// packages/features/account/account-profile-logic/client/model/types.ts
/*
Licensed Materials - Property of IBM
694906H
(c) Copyright IBM Corp.  2020 All Rights Reserved

US Government Users Restricted Rights - Use, duplication or disclosure restricted
by GSA ADP Schedule Contract with IBM Corp.
*/

export type Address = {
  id?: string;
  address1?: string;
  address2?: string;
  city?: string;
  company?: string;
  countryCode?: string;
  countryName?: string;
  firstName?: string;
  lastName?: string;
  name?: string;
  phone?: string;
  province?: string;
  zip?: string;
  email?: string;
  titleCode?: string;
};